# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI 模型提供商管理服务 API 文档\n提供模型提供商、模型信息和参数规则的查询接口",
        "title": "Genkit AI Service API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户列表，支持分页和租户过滤\n\n**权限要求**：\n- 平台管理员（system_admin）：可以查看所有租户的用户或指定租户的用户\n- 租户管理员（tenant_admin）：只能查看自己租户下的用户\n\n**tenantId 查询参数使用规则**：\n- 租户管理员：\n- tenantId 参数会被忽略\n- 始终只返回当前用户所属租户下的用户列表\n- 平台管理员：\n- 如果提供 tenantId，返回指定租户下的用户列表\n- 如果不提供 tenantId，返回所有租户下的用户列表\n\n**参数说明**：\n- pageNo: 页码（从1开始，默认1）\n- pageSize: 每页大小（1-100，默认20）\n- tenantId: 租户ID（可选，UUID格式，仅平台管理员可用）\n\n**注意事项**：\n- 租户管理员调用此接口时，tenantId 参数会被忽略\n- 租户管理员始终只能看到自己租户下的用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Management"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "pageNo",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "租户ID（仅平台管理员可用）",
                        "name": "tenantId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/UserListResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新用户，支持可选的租户ID参数\n\n**权限要求**：\n- 平台管理员（system_admin）：可以在任意租户下创建用户\n- 租户管理员（tenant_admin）：只能在自己的租户下创建用户\n\n**tenantId 参数使用规则**：\n- 租户管理员：\n- 如果不提供 tenantId，系统自动使用当前用户的租户ID\n- 如果提供 tenantId，必须与当前用户的租户ID匹配，否则返回 403 错误\n- 平台管理员：\n- 必须提供 tenantId 参数\n- 可以指定任意有效的租户ID\n\n**参数说明**：\n- tenantId: 租户ID（可选，UUID格式）\n- email: 用户邮箱（必填，需符合邮箱格式）\n- password: 用户密码（必填，最少8个字符）\n- displayName: 显示名称（可选）\n- phone: 手机号码（可选）\n- isAdmin: 是否为管理员（可选，默认false）\n- roles: 用户角色列表（可选）\n- meta: 用户元数据（可选，JSON对象）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Management"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "创建用户请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/UserDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足：租户管理员只能在当前租户下创建用户",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "启用或禁用用户\n\n**权限要求**：\n- 平台管理员（system_admin）：可以更新任意租户下的用户状态\n- 租户管理员（tenant_admin）：只能更新自己租户下的用户状态\n\n**访问权限验证**：\n- 系统会自动查询目标用户所属的租户\n- 租户管理员尝试更新其他租户的用户状态时，将收到 403 权限不足错误\n- 平台管理员可以更新任意租户的用户状态，无需额外验证\n\n**功能说明**：\n- 启用用户（isActive=true）：用户可以正常登录和访问系统\n- 禁用用户（isActive=false）：用户将无法登录系统\n\n**参数说明**：\n- id: 用户ID（UUID格式）\n- isActive: 用户状态（true=启用，false=禁用）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Management"
                ],
                "summary": "更新用户状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新用户状态请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateUserStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/UserDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足：无法更新其他租户的用户状态",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "displayName": {
                    "type": "string",
                    "example": "张三"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "isAdmin": {
                    "type": "boolean",
                    "example": false
                },
                "meta": {
                    "type": "object"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "password123"
                },
                "phone": {
                    "type": "string",
                    "example": "13800138000"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"user\"]"
                    ]
                },
                "tenantId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "UpdateUserStatusRequest": {
            "type": "object",
            "required": [
                "isActive"
            ],
            "properties": {
                "isActive": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "请求参数错误"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
...
        "PaginationData-array_model_User": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                },
                "pageNo": {
                    "description": "当前页码",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页大小",
                    "type": "integer",
                    "example": 10
                },
                "totalCount": {
                    "description": "总记录数",
                    "type": "integer",
                    "example": 100
                },
                "totalPage": {
                    "description": "总页数",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "UserDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/User"
                        }
                    ]
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
        "UserListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "分页数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/PaginationData-array_model_User"
                        }
                    ]
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {token}\" 格式的 JWT 令牌进行身份认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}